Pushing in some changes I made. Trying to set up Mongo right now.
Currently trying to figure out if I connected anything.

Current error when I "connected" it. everything works, everything is connected, but i get a weird error like the following =>

On my terminal (connection works fine)

This is the url list below this:
[]
[]
This is the url list below this:
[]
[]
This is the url list below this:
[]

Errors on web though. need to see whats going on.